requirements.txt
streamlit
streamlit-option-menu
opencv-python
cvzone
tensorflow
pillow
numpy
